+++ js/game-data.js
@@ -1,0 +1,72 @@
+export default [
+  {
+    type: 0,
+    answers: [
+      {
+        labels: [
+          {type: `photo`, text: `Фото`},
+          {type: `paint`, text: `Рисунок`}
+        ],
+        picture: {
+          src: `http://placehold.it/468x458`,
+          width: 468,
+          height: 458
+        }
+      },
+      {
+        labels: [
+          {type: `photo`, text: `Фото`},
+          {type: `paint`, text: `Рисунок`}
+        ],
+        picture: {
+          src: `http://placehold.it/468x458`,
+          width: 468,
+          height: 458
+        }
+      }
+    ]
+  }, {
+    type: 1,
+    answers: [
+      {
+        labels: [
+          {type: `photo`, text: `Фото`},
+          {type: `paint`, text: `Рисунок`}
+        ],
+        picture: {
+          src: `http://placehold.it/705x455`,
+          width: 705,
+          height: 455
+        }
+      }
+    ]
+  }, {
+    type: 2,
+    answers: [
+      {
+        labels: [],
+        picture: {
+          src: `http://placehold.it/304x455`,
+          width: 304,
+          height: 455
+        }
+      },
+      {
+        labels: [],
+        picture: {
+          src: `http://placehold.it/304x455`,
+          width: 304,
+          height: 455
+        }
+      },
+      {
+        labels: [],
+        picture: {
+          src: `http://placehold.it/304x455`,
+          width: 304,
+          height: 455
+        }
+      }
+    ]
+  }
+];
