--- js/game-three.js
@@ -1,53 +1,0 @@
-import addElementToPage from './add-element-to-page';
-import getElementFromTemplate from './get-element-from-template';
-import getGreeting from './greeting';
-import getResult from './result';
-import header from './header';
-import footer from './footer';
-
-
-const createScreen = () => {
-  const template = `
-  ${header}
-  <div class="game">
-    <p class="game__task">Найдите рисунок среди изображений</p>
-    <form class="game__content  game__content--triple">
-      <div class="game__option">
-        <img src="http://placehold.it/304x455" alt="Option 1" width="304" height="455">
-      </div>
-      <div class="game__option  game__option--selected">
-        <img src="http://placehold.it/304x455" alt="Option 1" width="304" height="455">
-      </div>
-      <div class="game__option">
-        <img src="http://placehold.it/304x455" alt="Option 1" width="304" height="455">
-      </div>
-    </form>
-    <div class="stats">
-      <ul class="stats">
-        <li class="stats__result stats__result--wrong"></li>
-        <li class="stats__result stats__result--slow"></li>
-        <li class="stats__result stats__result--fast"></li>
-        <li class="stats__result stats__result--correct"></li>
-        <li class="stats__result stats__result--wrong"></li>
-        <li class="stats__result stats__result--unknown"></li>
-        <li class="stats__result stats__result--slow"></li>
-        <li class="stats__result stats__result--unknown"></li>
-        <li class="stats__result stats__result--fast"></li>
-        <li class="stats__result stats__result--unknown"></li>
-      </ul>
-    </div>
-  </div>
-  ${footer}`;
-
-  const element = getElementFromTemplate(template);
-  const backButton = element.querySelector(`.back`);
-  const answers = [...element.querySelectorAll(`.game__option`)];
-
-  backButton.addEventListener(`click`, () => addElementToPage(getGreeting()));
-  answers.forEach((answer) => answer.addEventListener(`click`,
-      () => addElementToPage(getResult())));
-
-  return element;
-};
-
-export default createScreen;
