--- js/game-two.js
@@ -1,55 +1,0 @@
-import addElementToPage from './add-element-to-page';
-import getElementFromTemplate from './get-element-from-template';
-import getGreeting from './greeting';
-import getGameThree from './game-three';
-import header from './header';
-import footer from './footer';
-
-
-const createScreen = () => {
-  const template = `
-  ${header}
-  <div class="game">
-    <p class="game__task">Угадай, фото или рисунок?</p>
-    <form class="game__content  game__content--wide">
-      <div class="game__option">
-        <img src="http://placehold.it/705x455" alt="Option 1" width="705" height="455">
-        <label class="game__answer  game__answer--photo">
-          <input name="question1" type="radio" value="photo">
-          <span>Фото</span>
-        </label>
-        <label class="game__answer  game__answer--wide  game__answer--paint">
-          <input name="question1" type="radio" value="paint">
-          <span>Рисунок</span>
-        </label>
-      </div>
-    </form>
-    <div class="stats">
-      <ul class="stats">
-        <li class="stats__result stats__result--wrong"></li>
-        <li class="stats__result stats__result--slow"></li>
-        <li class="stats__result stats__result--fast"></li>
-        <li class="stats__result stats__result--correct"></li>
-        <li class="stats__result stats__result--wrong"></li>
-        <li class="stats__result stats__result--unknown"></li>
-        <li class="stats__result stats__result--slow"></li>
-        <li class="stats__result stats__result--unknown"></li>
-        <li class="stats__result stats__result--fast"></li>
-        <li class="stats__result stats__result--unknown"></li>
-      </ul>
-    </div>
-  </div>
-  ${footer}`;
-
-
-  const element = getElementFromTemplate(template);
-  const backButton = element.querySelector(`.back`);
-  const form = element.querySelector(`.game__content`);
-
-  backButton.addEventListener(`click`, () => addElementToPage(getGreeting()));
-  form.addEventListener(`change`, () => addElementToPage(getGameThree()));
-
-  return element;
-};
-
-export default createScreen;
