--- js/header/arrowPrev.js
@@ -1,6 +1,0 @@
-export default `<div class="header__back">
-    <span class="back">
-      <img src="img/arrow_left.svg" width="45" height="45" alt="Back">
-      <img src="img/logo_small.png" width="101" height="44">
-    </span>
-  </div>`;
