+++ js/templates/header/statistics.js
@@ -1,0 +1,6 @@
+export default `<h1 class="game__timer">NN</h1>
+<div class="game__lives">
+  <img src="img/heart__empty.svg" class="game__heart" alt="Life" width="32" height="32">
+  <img src="img/heart__full.svg" class="game__heart" alt="Life" width="32" height="32">
+  <img src="img/heart__full.svg" class="game__heart" alt="Life" width="32" height="32">
+</div>`;
