+++ js/templates/options/label.js
@@ -1,0 +1,6 @@
+export default (params) => {
+  return `<label class="game__answer game__answer--paint">
+      <input name="question1" type="radio" value="paint">
+      <span>Рисунок</span>
+    </label>`;
+};
