+++ js/templates/stats/game-stats.js
@@ -1,0 +1,13 @@
+export default (results) => `
+    <ul class="stats">
+      <li class="stats__result stats__result--wrong"></li>
+      <li class="stats__result stats__result--slow"></li>
+      <li class="stats__result stats__result--fast"></li>
+      <li class="stats__result stats__result--correct"></li>
+      <li class="stats__result stats__result--unknown"></li>
+      <li class="stats__result stats__result--unknown"></li>
+      <li class="stats__result stats__result--unknown"></li>
+      <li class="stats__result stats__result--unknown"></li>
+      <li class="stats__result stats__result--unknown"></li>
+      <li class="stats__result stats__result--unknown"></li>
+    </ul>`;
