+++ js/templates/stats/results.js
@@ -1,0 +1,2 @@
+export default (points, total) => `<td class="result__points">×&nbsp;${points}</td>
+<td class="result__total">${total}</td>`;
